+++ prboom2/src/dsda/map_format.h
@@ -53,7 +53,7 @@ typedef struct {
   void (*cross_special_line)(line_t *, int, mobj_t *, dboolean);
   void (*shoot_special_line)(mobj_t *, line_t *);
   dboolean (*test_activate_line)(line_t *, mobj_t *, int, line_activation_t);
-  dboolean (*execute_line_special)(int, byte *, line_t *, int, mobj_t *);
+  dboolean (*execute_line_special)(int, int *, line_t *, int, mobj_t *);
   void (*post_process_line_special)(line_t *);
   void (*post_process_sidedef_special)(side_t *, const mapsidedef_t *, sector_t *, int);
   void (*animate_surfaces)(void);
